Feat/generic computational graph avx (#215)
* start adding avx backend

* update avx backend implementation

* update avx_conv2d_back_kernel implementation

* trivial optimization

* optimize average_pooling_layer::forward_propagation, average_pooling_layer::back_propagation

* optimize layer::update_weight by replacing division with reciprocal multiplication
remove some warnings

* remove warnings

* update CMakeLists.txt (fix typo, squeeze out performance with MSVC build)

* rename avx_conv2d_back_kernel to avx_conv2d_back_5x5_kernel

* fix indentation

* change tabs to spaces

* fix issues pointed out by edgarriba and nyanp in #215

* separate kernel routines from class average_pooling_layer (#215)

* fix indentations #215

* fix compile errors with GCC #215

+++ examples/mnist/test_with_stb_image.cpp
@@ -50,25 +50,25 @@ void convert_image(const std::string& imagefilename,
     int w,
     int h,
     vec_t& data) {
-	// load
-	int input_w, input_h, comp;
-	stbi_uc* input_pixels = stbi_load(imagefilename.c_str(), &input_w, &input_h, &comp, 1);
-	if (!input_pixels) {
-		// cannot open, or it's not an image
-		cout << "stbi_load failed";
-		return;
-	}
+    // load
+    int input_w, input_h, comp;
+    stbi_uc* input_pixels = stbi_load(imagefilename.c_str(), &input_w, &input_h, &comp, 1);
+    if (!input_pixels) {
+        // cannot open, or it's not an image
+        cout << "stbi_load failed";
+        return;
+    }
 
-	// resize
-	std::vector<uint8_t> resized(w * h);
-	uint8_t* resized_pixels = &(resized[0]);
-	int input_stride_in_bytes = input_w;
-	if (!stbir_resize_uint8(input_pixels, input_w, input_h, input_stride_in_bytes, resized_pixels, w, h, w, 1)) {
-		cout << "stbir_resize_uint8 failed";
-		stbi_image_free(input_pixels);
-		return;
-	}
-	stbi_image_free(input_pixels);
+    // resize
+    std::vector<uint8_t> resized(w * h);
+    uint8_t* resized_pixels = &(resized[0]);
+    int input_stride_in_bytes = input_w;
+    if (!stbir_resize_uint8(input_pixels, input_w, input_h, input_stride_in_bytes, resized_pixels, w, h, w, 1)) {
+        cout << "stbir_resize_uint8 failed";
+        stbi_image_free(input_pixels);
+        return;
+    }
+    stbi_image_free(input_pixels);
 
     // mnist dataset is "white on black", so negate required
     std::transform(resized.begin(), resized.end(), std::back_inserter(data),
@@ -76,19 +76,19 @@ void convert_image(const std::string& imagefilename,
 }
 
 bool save_image(const std::string& imagefilename,
-	const image<>& img
-	)
+    const image<>& img
+    )
 {
-	// no scaling, save at original size
-	int stride_bytes = img.width();
-	int ret = stbi_write_png(
-		imagefilename.c_str(),
-		img.width(),
-		img.height(),
-		1,
-		&(img.at(0, 0)),
-		stride_bytes);
-	return (ret != 0);
+    // no scaling, save at original size
+    int stride_bytes = (int)img.width();
+    int ret = stbi_write_png(
+        imagefilename.c_str(),
+        (int)img.width(),
+        (int)img.height(),
+        1,
+        &(img.at(0, 0)),
+        stride_bytes);
+    return (ret != 0);
 }
 
 void construct_net(network<sequential>& nn) {
@@ -117,7 +117,7 @@ void construct_net(network<sequential>& nn) {
 }
 
 void recognize(const std::string& dictionary, const std::string& filename) {
-	network<sequential> nn;
+    network<sequential> nn;
 
     construct_net(nn);
 
@@ -145,19 +145,19 @@ void recognize(const std::string& dictionary, const std::string& filename) {
     // save outputs of each layer
     for (size_t i = 0; i < nn.depth(); i++) {
         auto out_img = nn[i]->output_to_image();
-		auto filename = "layer_" + std::to_string(i) + ".png";
-		if (!save_image(filename, out_img)) {
-			cout << "failed to save " << filename << endl;
-		}
+        auto filename = "layer_" + std::to_string(i) + ".png";
+        if (!save_image(filename, out_img)) {
+            cout << "failed to save " << filename << endl;
+        }
     }
     // save filter shape of first convolutional layer
-	{
-	    auto weight = nn.at<convolutional_layer<tan_h>>(0).weight_to_image();
-		auto filename = "weights.png";
-		if (!save_image(filename, weight)) {
-			cout << "failed to save " << filename << endl;
-		}
-	}
+    {
+        auto weight = nn.at<convolutional_layer<tan_h>>(0).weight_to_image();
+        auto filename = "weights.png";
+        if (!save_image(filename, weight)) {
+            cout << "failed to save " << filename << endl;
+        }
+    }
 }
 
 int main(int argc, char** argv) {
